Author: Thomas Ruehr <ruehr@in.tum.de> Date:   Fri Jul 29 14:36:16 2011 +0200
Helper node to determine a base pose that makes executing cartesian trajectories with the arms feasible. 
The action interface should be straight forward, the input is a vector of positions in arbitrary frame together with a vector of arm indices, 0 for right and 1 for left arm. 
The result will be a pose in base_link that makes the trajectories for both arms reachable. Currently only poses are generated that do not bring the base into collision with base_laser obstacles. 
This and other features will be pulled out as parameters in the future.
Use with caution, some bugs might be hidden here and there ;)

git-commit